Move and split some types into their own files.
